refactor(runners): extract grok turn runner to runners/grok.rs (Phase 2a)
run_grok_turn, the grok CLI bootstrap/auth-sync helpers, the grok event
parsers, and the interactive-login detectors move verbatim. Shared
helpers they use (command_available, usage_value_tokens,
merge_stream_fragment, TextDeltaCoalescer, thinking_final_event,
turn-outcome/cost plumbing) become pub(crate) in mission_runner.
mission_runner re-exports the moved symbols so control.rs dispatch and
tests keep their paths. No behavior change.
